Add encouraging text to chat intros
