+++ mobile/src/components/chat/utils.ts
@@ -88,3 +88,16 @@ export function shouldHighlightChatRoom(
 
   return true;
 }
+
+export function makeListSentence(words: string[]): string {
+  if (words.length === 0) {
+    return "";
+  }
+  if (words.length === 1) {
+    return words[0];
+  }
+  if (words.length === 2) {
+    return words.join(" and ");
+  }
+  return words.slice(0, -1).join(", ") + ", and " + words[words.length - 1];
+}
