+++ web/components/chats/RenderChatRoom.tsx
@@ -1,24 +1,25 @@
 import { useCallback } from "react";
 
-import { useDisclosure } from "@mantine/hooks";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import toast from "react-hot-toast";
 
+import { makeListSentence } from "../../common/lib/words";
 import {
   Chat_Message,
   User_Type_Enum,
   useSendMessageMutation,
 } from "../../generated/graphql";
-import { BxChevronLeft, BxInfoCircle } from "../../generated/icons/regular";
+import { BxChevronLeft } from "../../generated/icons/regular";
 import { useCurrentProfile } from "../../hooks/useCurrentProfile";
 import { useMediaQuery } from "../../hooks/useMediaQuery";
 import { useQueryParam } from "../../hooks/useQueryParam";
 import { PromiseQueue } from "../../lib/PromiseQueue";
+import { getFirstNameOfUser } from "../../lib/user";
 import { Text } from "../atomic";
 import { ProfileImage } from "../common/ProfileImage";
-import { ActionModal } from "../modals/ActionModal";
 
+import { ChatParticipantsModalButton } from "./ChatParticipantsModalButton";
 import { ChatRoomImage } from "./ChatRoomImage";
 import { ChatTitle } from "./ChatTitle";
 import { RenderChatRoomMessages } from "./RenderChatRoomMessages";
@@ -115,75 +116,10 @@ export function RenderChatRoom() {
       </div>
 
       <RenderChatRoomMessages chatRoomId={chatRoomId ?? ""} />
+
       <div className="h-px w-full shrink-0 bg-gray-600"></div>
 
       <SendMessageInput className="m-4 ml-16" chatRoomId={chatRoomId ?? null} />
     </div>
-  );
-}
-
-interface ChatParticipantsModalButtonProps {
-  chatParticipants: ChatParticipant[];
-}
-function ChatParticipantsModalButton(props: ChatParticipantsModalButtonProps) {
-  const { chatParticipants } = props;
-
-  const { currentProfile } = useCurrentProfile();
-  const spaceSlug = useQueryParam("slug", "string");
-  const [open, handlers] = useDisclosure(false);
-
-  return (
-    <>
-      <BxInfoCircle
-        className="h-5 w-5 cursor-pointer text-gray-500"
-        onClick={handlers.open}
-      />
-
-      <ActionModal
-        isOpen={open}
-        onClose={handlers.close}
-        actionText="Close"
-        onAction={handlers.close}
-      >
-        <div className="flex flex-col rounded-md bg-white px-8 py-8">
-          <div className="w-96"></div>
-          <Text variant="subheading1" className="mx-auto">
-            People in this chat
-          </Text>
-          <div className="h-8"></div>
-          <div className="flex flex-col gap-3">
-            {chatParticipants.map((p) => (
-              <div
-                className="flex items-center justify-between"
-                key={p.profileId}
-              >
-                <div className="flex items-center gap-2">
-                  <ProfileImage
-                    className="h-10 w-10"
-                    src={p.profileImage?.url}
-                  />
-                  <Text variant="body1">{p.fullName}</Text>
-                </div>
-                {p.profileId === currentProfile?.id ? (
-                  <Text variant="body1" className="text-gray-500">
-                    You
-                  </Text>
-                ) : (
-                  <a
-                    href={`/space/${spaceSlug}/profile/${p.profileId}`}
-                    target="_blank"
-                    rel="noreferrer"
-                  >
-                    <Text className="text-gray-500 hover:underline">
-                      View Profile
-                    </Text>
-                  </a>
-                )}
-              </div>
-            ))}
-          </div>
-        </div>
-      </ActionModal>
-    </>
   );
 }
