+++ web/components/chats/RenderChatRoomMessages.tsx
@@ -1,5 +1,7 @@
 import { useState, useEffect, useMemo } from "react";
 
+import { makeListSentence } from "../../common/lib/words";
+import { User_Type_Enum } from "../../generated/graphql";
 import { BxSend } from "../../generated/icons/regular";
 import { useCurrentProfile } from "../../hooks/useCurrentProfile";
 import { usePrevious } from "../../hooks/usePrevious";
@@ -11,6 +13,7 @@ import { RenderMessage } from "./RenderMessage";
 import { SendMessageInput } from "./SendMessageInput";
 import { useChatRoom } from "./useChatRoom";
 import { useMessages } from "./useMessages";
+import { getChatParticipants } from "./utils";
 
 interface RenderChatRoomMessagesProps {
   chatRoomId: string;
@@ -19,7 +22,7 @@ interface RenderChatRoomMessagesProps {
 export function RenderChatRoomMessages(props: RenderChatRoomMessagesProps) {
   const { chatRoomId } = props;
 
-  const { chatParticipants } = useChatRoom(chatRoomId ?? "");
+  const { chatParticipants, chatRoom } = useChatRoom(chatRoomId ?? "");
 
   const { currentProfile } = useCurrentProfile();
   const {
@@ -47,6 +50,21 @@ export function RenderChatRoomMessages(props: RenderChatRoomMessagesProps) {
 
   const prevLastMessageIdByOther = usePrevious(lastMessageIdByOther);
 
+  const isIntro = !!chatRoom?.chat_intro_id;
+
+  // Chat topbar
+  const allHumans = chatParticipants.filter(
+    (p) => p.userType !== User_Type_Enum.Bot
+  );
+  const otherHumans = allHumans.filter(
+    (p) => p.profileId !== currentProfile?.id
+  );
+
+  const otherHumanNames =
+    otherHumans.length <= 2
+      ? makeListSentence(otherHumans.map((p) => p.firstName))
+      : "the others";
+
   useEffect(() => {
     if (document.hidden) {
       return;
@@ -70,6 +88,16 @@ export function RenderChatRoomMessages(props: RenderChatRoomMessagesProps) {
   return (
     <div className="flex w-full flex-1 flex-col overflow-hidden">
       <div className="flex flex-1 flex-col-reverse overflow-y-scroll overscroll-contain p-4">
+        {isIntro && (
+          <div className="flex w-full flex-col items-center">
+            <Text className="text-gray-400" variant="body3">
+              Introduce yourself to {otherHumanNames}!
+            </Text>
+            <Text className="text-gray-400" variant="body3">
+              Remember: Everyone on Canopy is here to meet people.
+            </Text>
+          </div>
+        )}
         {currentProfile &&
           messagesList.map((message, idx) => {
             // Note: Messages are ordered by created_at DESC
